chore(pipeline): set up TSLint in pipeline closes #13
As explained in #13 with this addition lint runs automatically in CI pipeline. I've also reorganised the folder structure to bring it in line with the API repository. Next supports a flat structure out of the box, but since Web uses custom server, it was possible to move source files to `./src`.