chore(pipeline): set up TSLint in pipeline closes #13 (#14)
* chore(pipeline): set up TSLint in pipeline closes #13

As explained in #13 with this addition lint runs automatically in CI pipeline. I've also reorganised the folder structure to bring it in line with the API repository. Next supports a flat structure out of the box, but since Web uses custom server, it was possible to move source files to `./src`.

* chore(ci): revert to using default next build and serve

The config was getting messy so I cleaned it up in favour of a future release of Next and Now supporting serving icons from `/static` at root.

* chore(ci): remove unnecessary build ID generation

It was not in the scope of this.

* chore(ci): changed the config src in now.json to point to package.json

* fix(now): set the output dir relative to next.config.js

This was required for Now.sh to actually see our built pages.

* fix(now): replace __dirname with relative path

Now builder parses distDir in a way that resolves relatively. This is not good, so a relative `../.next` had to be used instead of __dirname.